app/components/nav-bar/nav-bar.tsx
import { SignedIn, SignedOut, SignInButton, UserButton } from "@clerk/nextjs";

export default function NavBar() {
  return (
    <header className="flex justify-between items-center p-4 bg-gray-800 text-white">
      <h1 className="text-2xl font-semibold">Script Caddy</h1>
      <nav>
        <ul className="flex gap-4">
          <li>
            <SignedOut>
              <SignInButton />
            </SignedOut>
          </li>
          <li>
            <SignedIn>
              <UserButton />
            </SignedIn>
          </li>
        </ul>
      </nav>
    </header>
  );
}
